Applied Design Patterns for Mines
Separated code into different classes and also new javascript file that takes control of the game
